chore: refactor folder structure to be more go-like
