+++ chat-service/cmd/main.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	"workout-planner/chat/config"
+)
+
+// initial implementation borrowed from https://gowebexamples.com/websockets/
+func main() {
+	config.StartServer()
+}
