+++ chat-service/go.mod
@@ -2,7 +2,6 @@ module workout-planner/chat
 
 go 1.22.1
 
-require (
-	github.com/gorilla/websocket v1.5.1 // indirect
-	golang.org/x/net v0.17.0 // indirect
-)
+require github.com/gorilla/websocket v1.5.1
+
+require golang.org/x/net v0.17.0 // indirect
