+++ chat-service/static/client2.html
